Fix 日付ごとの日記取得処理の実装
