+++ front/src/app/(authenticated)/diaries/[date]/page.tsx
@@ -1,0 +1,8 @@
+
+export default function Page({ params }: { params: { date: string } }) {
+  return (
+    <div className="mx-auto ml-5 w-full">
+      {params.date}
+    </div>
+  )
+}
