--- front/src/app/(authenticated)/diaries/[uid]/page.tsx
@@ -1,7 +1,0 @@
-
-export default function Page() {
-  return (
-    <div className="mx-auto ml-5 w-full">
-    </div>
-  )
-}
