+++ front/src/features/diaries/api/index.ts
@@ -1,2 +1,3 @@
 export * from './create-diary'
 export * from './get-diaries'
+export * from './get-dairy-diaries'
